+++ inherited-rules/implicitly-included/jsdoc-recommended.js
@@ -4,14 +4,10 @@ module.exports = {
     "jsdoc/check-property-names": "warn",
     "jsdoc/check-values": "warn",
     "jsdoc/empty-tags": "warn",
-    "jsdoc/implements-on-classes": "warn",
     "jsdoc/require-param": "warn",
-    "jsdoc/require-param-name": "warn",
-    "jsdoc/require-param-type": "warn",
     "jsdoc/require-property": "warn",
     "jsdoc/require-property-name": "warn",
     "jsdoc/require-property-type": "warn",
-    "jsdoc/require-returns-check": "warn",
-    "jsdoc/require-returns-type": "warn"
+    "jsdoc/require-returns-check": "warn"
   }
 };